requirements.txt
fastapi
uvicorn[standard]
langchain
langgraph
langchain-core
pydantic
python-dotenv
httpx
pydantic-settings
langchain-google-genai
langmem
mem0ai
pinecone-text
langchain-pinecone